+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Patch] Add a `Button` example that has an icon within it.
+
 ### Changed
 - [Feature] `OverlayWrapper` prop `onHide` can now prevent the `overlay` from hiding by returning false. (#664)